+++ src/collective/z3cform/colorpicker/__init__.py
@@ -1,0 +1,4 @@
+from .colorpicker import ColorpickerFieldWidget
+from .colorpickeralpha import ColorpickerAlphaFieldWidget
+from .fields import Color
+from .fields import ColorAlpha