+++ src/collective/z3cform/colorpicker/demo/browser.py
@@ -1,44 +1,42 @@
-from zope.interface import implements
-from zope.interface import Interface
-from z3c.form import form, field, button
-from zope import schema
 from plone.z3cform.layout import wrap_form
-from ..colorpicker import ColorpickerFieldWidget
-from ..colorpickeralpha import ColorpickerAlphaFieldWidget
+from z3c.form import form, field, button
+from zope.interface import Interface
+from collective.z3cform.colorpicker import Color
+from collective.z3cform.colorpicker import ColorAlpha
 
 
 class IColorForm(Interface):
-    color = schema.TextLine(title=u"Color",
-                               description=u"",
-                               required=False)
+    color = Color(
+        title=u"Color",
+        description=u"",
+        required=False,
+        default=u"#6298c9")
 
-    alphacolor = schema.TextLine(title=u"Color with alpha layer support",
-                               description=u"",
-                               required=False)
-
-
-class Color(object):
-    implements(IColorForm)
-    color = '#ff0000'
-    alphacolor = 'ff0000cc'
-
-    def __init__(self, context):
-        self.context = context
+    alphacolor = ColorAlpha(
+        title=u"Color with alpha layer support",
+        description=u"",
+        required=False,
+        default=u"rgba(104,191,144,0.55)")
 
 
 class ColorDemoForm(form.Form):
     """Example color picker form"""
 
     fields = field.Fields(IColorForm)
-    fields['color'].widgetFactory = ColorpickerFieldWidget
-    fields['alphacolor'].widgetFactory = ColorpickerAlphaFieldWidget
+    ignoreContext = True
+
+    label = u"Colopicker Demo"
+    description = (
+        u"Color picker widget based on "
+        u"http://mjolnic.com/bootstrap-colorpicker/"
+    )
 
     def __init__(self, context, request):
         super(ColorDemoForm, self).__init__(context, request)
         self.request.set('disable_border', True)
 
     @button.buttonAndHandler(u'Save')
-    def handleApply(self, action):  # pylint: disable=W0613
+    def handleApply(self, action):
         # data, errors = self.extractData()
         return